backend/requirements.txt
bcrypt==4.0.1
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.121.0
passlib==1.7.4
python-jose==3.5.0
sqlalchemy==2.0.44
uvicorn==0.38.0
asyncpg==0.30.0
httpx==0.28.1
setuptools
prometheus-fastapi-instrumentator==7.0.2
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0